+++ src/Components/Movie/Single Movie/SingleMovie.js
@@ -3,7 +3,7 @@ import { useParams } from "react-router-dom";
 function SingleMovie() {
   const { id } = useParams();
 
-  return <div className="single-movie">single movie! {id}</div>;
+  return <div className="single-movie">single movie!{id}</div>;
 }
 
 export default SingleMovie;
